Moved LICENSE back, modified Makefile.

+++ trunk/Makefile
@@ -1,5 +1,6 @@
 # VARIABLES
 
+USER=$(shell whoami)
 CC = g++
 CFLAGS = -Wall -Os
 D_CFLAGS = -Wall -D _DEBUG
@@ -7,46 +8,46 @@ LDFLAGS = -lm
 OBJECTS = obj/main.o obj/Assembler.o obj/Error.o obj/crc.o
 D_OBJECTS = obj/main.d.o obj/Assembler.d.o obj/Error.d.o obj/crc.d.o
 
-.PHONY: all debug clean
+.PHONY: all debug clean install uninstall
 
 #####################################################################
 # RELEASE TARGET (DEFAULT)
 
-tchip16: ${OBJECTS}
-	${CC} ${CFLAGS} ${OBJECTS} ${LDFLAGS} -s -o $@
+tchip16: $(OBJECTS)
+	$(CC) $(CFLAGS) $(OBJECTS) $(LDFLAGS) -s -o $@
 
 obj/main.o: src/main.cpp src/Error.h src/Assembler.h
-	${CC} -c ${CFLAGS} src/main.cpp -o $@ 
+	$(CC) -c $(CFLAGS) src/main.cpp -o $@ 
 
 obj/Assembler.o: src/Assembler.cpp src/Assembler.h src/Opcodes.h src/RomHeader.h src/crc.h
-	${CC} -c ${CFLAGS} src/Assembler.cpp -o $@
+	$(CC) -c $(CFLAGS) src/Assembler.cpp -o $@
 
 obj/Error.o: src/Error.cpp src/Error.h
-	${CC} -c ${CFLAGS} src/Error.cpp -o $@ 
+	$(CC) -c $(CFLAGS) src/Error.cpp -o $@ 
 
 obj/crc.o: src/crc.c src/crc.h
-	${CC} -c ${CFLAGS} src/crc.c -o $@
+	$(CC) -c $(CFLAGS) src/crc.c -o $@
 
 # DEBUG TARGET
 
 debug: tchip16_debug
 	
-tchip16_debug: ${D_OBJECTS}
-	${CC} ${D_CFLAGS} ${D_OBJECTS} ${LDFLAGS} -o $@
+tchip16_debug: $(D_OBJECTS)
+	$(CC) $(D_CFLAGS) $(D_OBJECTS) $(LDFLAGS) -o $@
 
 # DEBUG OBJECTS
 
 obj/main.d.o: src/main.cpp src/Error.h src/Assembler.h
-	${CC} -c ${D_CFLAGS} src/main.cpp -o $@ 
+	$(CC) -c $(D_CFLAGS) src/main.cpp -o $@ 
 
 obj/Assembler.d.o: src/Assembler.cpp src/Assembler.h src/Opcodes.h
-	${CC} -c ${D_CFLAGS} src/Assembler.cpp -o $@ 
+	$(CC) -c $(D_CFLAGS) src/Assembler.cpp -o $@ 
 
 obj/Error.d.o: src/Error.cpp src/Error.h
-	${CC} -c ${D_CFLAGS} src/Error.cpp -o $@ 
+	$(CC) -c $(D_CFLAGS) src/Error.cpp -o $@ 
 
 obj/crc.d.o: src/crc.c src/crc.h
-	${CC} -c ${D_CFLAGS} src/crc.c -o $@ 
+	$(CC) -c $(D_CFLAGS) src/crc.c -o $@ 
 
 #####################################################################
 # ALL TARGETS
@@ -61,19 +62,27 @@ clean:
 # (UN)INSTALL TARGET
 
 install:
+ifneq ($(USER),root)
+	-@echo You are not root, please run: \'sudo make install\'
+else
 	-@echo Installing binaries...
 	-@cp tchip16 tchip16_debug /usr/bin
 	-@echo Installing man page... 
 	-@cp src/tchip16.1.gz /usr/share/man/man1/tchip16.1.gz
 	-@echo Updating man-db...
-	-@mandb -q #> /dev/null || true
+	-@mandb -q
 	-@echo done.
+endif
 
 uninstall:
+ifneq ($(USER),root)
+	-@echo You are not root, please run: \'sudo make uninstall\'
+else
 	-@echo Removing binaries...
 	-@rm /usr/bin/tchip16 /usr/bin/tchip16_debug 2> /dev/null || true
 	-@echo Removing man page...
 	-@rm /usr/share/man/man1/tchip16.1.gz 2> /dev/null || true
 	-@echo Updating man-db... 
-	-@mandb -q #> /dev/null || true
+	-@mandb -q 
 	-@echo done.
+endif
